Attempt at reconfigurable config service as a friday project; failed due
to explosion in complexity. Have a better approach in mind now; the main
problem was allowing a leader/server operate in multiple configs. The
SMART (Lorch et al.) approach is to have configuration-specific
replicas. This would get rid of a lot of the complexity.

Also check in attempts at making pb generic over the state-migration
plan, also failed because stuff is too intertwined.